Remove duplication by extracting dynamic sections to a component
