+++ frontend/src/employee-frontend/components/vasu/VasuEditPage.tsx
@@ -3,43 +3,25 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 import Button from 'lib-components/atoms/buttons/Button'
-import { cloneDeep } from 'lodash'
-import React, { Fragment } from 'react'
+import React from 'react'
 import { useHistory } from 'react-router'
 import { RouteComponentProps } from 'react-router-dom'
 import styled from 'styled-components'
 import { DATE_FORMAT_TIME_ONLY, formatDate } from '../../../lib-common/date'
 import { UUID } from '../../../lib-common/types'
 import Spinner from '../../../lib-components/atoms/state/Spinner'
-import {
-  Container,
-  ContentArea
-} from '../../../lib-components/layout/Container'
+import { Container } from '../../../lib-components/layout/Container'
 import StickyFooter from '../../../lib-components/layout/StickyFooter'
-import { Dimmed, H2 } from '../../../lib-components/typography'
+import { Dimmed } from '../../../lib-components/typography'
 import { defaultMargins, Gap } from '../../../lib-components/white-space'
 import { useTranslation } from '../../state/i18n'
-import { CheckboxQuestion as CheckboxQuestionElem } from './components/CheckboxQuestion'
-import { MultiSelectQuestion as MultiSelectQuestionElem } from './components/MultiSelectQuestion'
-import { RadioGroupQuestion as RadioGroupQuestionElem } from './components/RadioGroupQuestion'
-import { TextQuestion as TextQuestionElem } from './components/TextQuestion'
 import { EditableAuthorsSection } from './sections/AuthorsSection'
+import { DynamicSections } from './sections/DynamicSections'
 import { EditableEvaluationDiscussionSection } from './sections/EvaluationDiscussionSection'
 import { EditableVasuDiscussionSection } from './sections/VasuDiscussionSection'
 import { VasuEvents } from './sections/VasuEvents'
 import { VasuHeader } from './sections/VasuHeader'
 import { useVasu, VasuStatus } from './use-vasu'
-import {
-  CheckboxQuestion,
-  isCheckboxQuestion,
-  isMultiSelectQuestion,
-  isRadioGroupQuestion,
-  isTextQuestion,
-  MultiSelectQuestion,
-  QuestionOption,
-  RadioGroupQuestion,
-  TextQuestion
-} from './vasu-content'
 
 const FooterContainer = styled.div`
   display: flex;
@@ -102,10 +84,6 @@ export default React.memo(function VasuEditPage({
   const showSpinner = status.state === 'saving'
 
   const dynamicSectionsOffset = 1
-  const getDynamicQuestionNumber = (
-    sectionIndex: number,
-    questionIndex: number
-  ) => `${sectionIndex + 1 + dynamicSectionsOffset}.${questionIndex + 1}`
 
   return (
     <Container>
@@ -120,108 +98,11 @@ export default React.memo(function VasuEditPage({
             setContent={setAuthorsContent}
           />
           <Gap size={'L'} />
-          {content.sections.map((section, sectionIndex) => {
-            return (
-              <Fragment key={section.name}>
-                <ContentArea opaque>
-                  <H2>
-                    {sectionIndex + 1 + dynamicSectionsOffset}. {section.name}
-                  </H2>
-                  {section.questions.map((question, questionIndex) => {
-                    const questionNumber = getDynamicQuestionNumber(
-                      sectionIndex,
-                      questionIndex
-                    )
-                    return (
-                      <Fragment key={question.name}>
-                        {isTextQuestion(question) ? (
-                          <TextQuestionElem
-                            question={question}
-                            questionNumber={questionNumber}
-                            onChange={(value) =>
-                              setContent((prev) => {
-                                const clone = cloneDeep(prev)
-                                const question1 = clone.sections[sectionIndex]
-                                  .questions[questionIndex] as TextQuestion
-                                question1.value = value
-                                return clone
-                              })
-                            }
-                          />
-                        ) : isCheckboxQuestion(question) ? (
-                          <CheckboxQuestionElem
-                            question={question}
-                            questionNumber={questionNumber}
-                            onChange={(checked) =>
-                              setContent((prev) => {
-                                const clone = cloneDeep(prev)
-                                const question1 = clone.sections[sectionIndex]
-                                  .questions[questionIndex] as CheckboxQuestion
-                                question1.value = checked
-                                return clone
-                              })
-                            }
-                          />
-                        ) : isRadioGroupQuestion(question) ? (
-                          <RadioGroupQuestionElem
-                            questionNumber={questionNumber}
-                            question={question}
-                            selectedValue={
-                              (content.sections[sectionIndex].questions[
-                                questionIndex
-                              ] as RadioGroupQuestion).value
-                            }
-                            onChange={(selectedOption: QuestionOption) =>
-                              setContent((prev) => {
-                                const clone = cloneDeep(prev)
-                                const question1 = clone.sections[sectionIndex]
-                                  .questions[
-                                  questionIndex
-                                ] as RadioGroupQuestion
-                                question1.value = selectedOption.key
-                                return clone
-                              })
-                            }
-                          />
-                        ) : isMultiSelectQuestion(question) ? (
-                          <MultiSelectQuestionElem
-                            question={question}
-                            questionNumber={questionNumber}
-                            selectedValues={
-                              (content.sections[sectionIndex].questions[
-                                questionIndex
-                              ] as MultiSelectQuestion).value
-                            }
-                            onChange={(option, checked) =>
-                              setContent((prev) => {
-                                const clone = cloneDeep(prev)
-                                const question1 = clone.sections[sectionIndex]
-                                  .questions[
-                                  questionIndex
-                                ] as MultiSelectQuestion
-                                if (
-                                  checked &&
-                                  !question1.value.includes(option.key)
-                                )
-                                  question1.value.push(option.key)
-                                if (!checked)
-                                  question1.value = question1.value.filter(
-                                    (i) => i !== option.key
-                                  )
-                                return clone
-                              })
-                            }
-                          />
-                        ) : undefined}
-                        <Gap size={'L'} />
-                      </Fragment>
-                    )
-                  })}
-                </ContentArea>
-                <Gap size={'L'} />
-              </Fragment>
-            )
-          })}
+          <DynamicSections
+            sectionIndex={dynamicSectionsOffset}
+            sections={content.sections}
+            setContent={setContent}
+          />
           <EditableVasuDiscussionSection
             sectionIndex={content.sections.length + dynamicSectionsOffset}
             content={vasuDiscussionContent}
